Refactored and fixed Python implementation
